+++ api/src/prescription/campaign/application/campaign-detail-controller.js
@@ -51,6 +51,38 @@ const findPaginatedFilteredCampaigns = async function (
   return dependencies.campaignReportSerializer.serialize(campaigns, meta);
 };
 
+const getCsvAssessmentResults = async function (request, h, dependencies = { tokenService }) {
+  const token = request.query.accessToken;
+  const { userId, campaignId: extractedCampaignId } =
+    dependencies.tokenService.extractCampaignResultsTokenContent(token);
+  const campaignId = request.params.id;
+
+  if (extractedCampaignId !== campaignId) {
+    throw new ForbiddenAccess();
+  }
+
+  const writableStream = new PassThrough();
+
+  const { fileName } = await usecases.startWritingCampaignAssessmentResultsToStream({
+    userId,
+    campaignId,
+    writableStream,
+    i18n: request.i18n,
+  });
+  const escapedFileName = escapeFileName(fileName);
+
+  writableStream.headers = {
+    'content-type': 'text/csv;charset=utf-8',
+    'content-disposition': `attachment; filename="${escapedFileName}"`,
+
+    // WHY: to avoid compression because when compressing, the server buffers
+    // for too long causing a response timeout.
+    'content-encoding': 'identity',
+  };
+
+  return writableStream;
+};
+
 const getCsvProfilesCollectionResults = async function (request, h, dependencies = { tokenService }) {
   const token = request.query.accessToken;
   const { userId, campaignId: extractedCampaignId } =
@@ -86,6 +118,7 @@ const getCsvProfilesCollectionResults = async function (request, h, dependencies
 const campaignDetailController = {
   getById,
   findPaginatedFilteredCampaigns,
+  getCsvAssessmentResults,
   getCsvProfilesCollectionResults,
 };
