+++ api/src/prescription/campaign/infrastructure/repositories/campaign-participation-info-repository.js
@@ -1,6 +1,6 @@
-import { knex } from '../../../db/knex-database-connection.js';
-import { Assessment } from '../../../src/shared/domain/models/Assessment.js';
-import { CampaignParticipationInfo } from '../../domain/read-models/CampaignParticipationInfo.js';
+import { knex } from '../../../../../db/knex-database-connection.js';
+import { Assessment } from '../../../../shared/domain/models/Assessment.js';
+import { CampaignParticipationInfo } from '../../../../../lib/domain/read-models/CampaignParticipationInfo.js';
 
 const findByCampaignId = async function (campaignId) {
   const results = await knex
