+++ api/tests/acceptance/application/campaigns/campaign-controller_test.js
@@ -1,5 +1,3 @@
-import jwt from 'jsonwebtoken';
-
 import { CampaignParticipationStatuses } from '../../../../src/prescription/shared/domain/constants.js';
 import {
   databaseBuilder,
@@ -9,7 +7,6 @@ import {
   mockLearningContent,
 } from '../../../test-helper.js';
 
-import { config as settings } from '../../../../lib/config.js';
 import { Membership } from '../../../../lib/domain/models/Membership.js';
 import { createServer } from '../../../../server.js';
 
@@ -184,96 +181,6 @@ describe('Acceptance | API | Campaign Controller', function () {
       // then
       expect(response.statusCode).to.equal(200, response.payload);
       expect(response.result).to.deep.equal(expectedResult);
-    });
-  });
-
-  describe('GET /api/campaigns/{id}/csv-assessment-results', function () {
-    let accessToken;
-
-    function _createTokenWithAccessId({ userId, campaignId }) {
-      return jwt.sign(
-        {
-          access_id: userId,
-          campaign_id: campaignId,
-        },
-        settings.authentication.secret,
-        { expiresIn: settings.authentication.accessTokenLifespanMs },
-      );
-    }
-
-    beforeEach(async function () {
-      const skillId = 'rec123';
-      const userId = databaseBuilder.factory.buildUser().id;
-      organization = databaseBuilder.factory.buildOrganization();
-      campaign = databaseBuilder.factory.buildCampaign({
-        organizationId: organization.id,
-      });
-      databaseBuilder.factory.buildCampaignSkill({ campaignId: campaign.id, skillId: skillId });
-      accessToken = _createTokenWithAccessId({ userId, campaignId: campaign.id });
-
-      databaseBuilder.factory.buildMembership({
-        userId,
-        organizationId: organization.id,
-        organizationRole: Membership.roles.MEMBER,
-      });
-
-      await databaseBuilder.commit();
-
-      const learningContent = [
-        {
-          id: 'recArea1',
-          title_i18n: {
-            fr: 'area1_Title',
-          },
-          color: 'specialColor',
-          competences: [
-            {
-              id: 'recCompetence1',
-              name_i18n: { fr: 'Fabriquer un meuble' },
-              index: '1.1',
-              tubes: [
-                {
-                  id: 'recTube1',
-                  skills: [
-                    {
-                      id: skillId,
-                      nom: '@web2',
-                      challenges: [],
-                    },
-                  ],
-                },
-              ],
-            },
-          ],
-        },
-      ];
-
-      const learningContentObjects = learningContentBuilder.fromAreas(learningContent);
-      mockLearningContent(learningContentObjects);
-    });
-
-    it('should return csv file with statusCode 200', async function () {
-      // given & when
-      const { statusCode, payload } = await server.inject({
-        method: 'GET',
-        url: `/api/campaigns/${campaign.id}/csv-assessment-results?accessToken=${accessToken}`,
-      });
-
-      // then
-      expect(statusCode).to.equal(200, payload);
-    });
-
-    context('when accessing another campaign with the wrong access token', function () {
-      it('should return an error response with an HTTP status code 403', async function () {
-        // given & when
-        const { statusCode } = await server.inject({
-          method: 'GET',
-          url: `/api/campaigns/1234567890/csv-assessment-results?accessToken=${accessToken}`,
-        });
-
-        // then
-        expect(statusCode).to.equal(403);
-      });
     });
   });
