+++ api/tests/integration/application/campaigns/index_test.js
@@ -6,22 +6,10 @@ describe('Integration | Application | Route | campaignRouter', function () {
   let httpTestServer;
 
   beforeEach(async function () {
-    sinon.stub(campaignController, 'getCsvAssessmentResults').callsFake((request, h) => h.response('ok').code(200));
-
-    sinon.stub(campaignController, 'getAnalysis').callsFake((request, h) => h.response('ok').code(200));
+    sinon.stub(campaignController, 'getAnalysis').callsFake((_, h) => h.response('ok').code(200));
 
     httpTestServer = new HttpTestServer();
     await httpTestServer.register(moduleUnderTest);
-  });
-
-  describe('GET /api/campaigns/{id}/csv-assessment-results', function () {
-    it('should exist', async function () {
-      // when
-      const response = await httpTestServer.request('GET', '/api/campaigns/1/csv-assessment-results');
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
   });
 
   describe('GET /api/campaigns/{id}/analyses', function () {
