+++ api/tests/prescription/campaign/integration/application/campaign-detail-route_test.js
@@ -7,11 +7,42 @@ describe('Integration | Application | Route | campaign detail router', function 
   let httpTestServer;
 
   beforeEach(async function () {
+    sinon.stub(campaignDetailController, 'getCsvAssessmentResults').callsFake((_, h) => h.response('ok').code(200));
     sinon
       .stub(campaignDetailController, 'getCsvProfilesCollectionResults')
       .callsFake((_, h) => h.response('ok').code(200));
+
     httpTestServer = new HttpTestServer();
     await httpTestServer.register(moduleUnderTest);
+  });
+
+  describe('GET /api/campaigns/{id}/csv-profiles-collection-results', function () {
+    it('should exist', async function () {
+      // given
+      const method = 'GET';
+      const url = '/api/campaigns/1/csv-profiles-collection-results';
+
+      // when
+      const response = await httpTestServer.request(method, url);
+
+      // then
+      expect(response.statusCode).to.equal(200);
+    });
+  });
+
+  describe('GET /api/campaigns/{id}/csv-assessment-results', function () {
+    it('should exist', async function () {
+      // given
+      const method = 'GET';
+      const url = '/api/campaigns/1/csv-assessment-results';
+
+      // when
+      const response = await httpTestServer.request(method, url);
+
+      // then
+      expect(response.statusCode).to.equal(200);
+      expect(campaignDetailController.getCsvAssessmentResults).to.have.been.calledOnce;
+    });
   });
 
   describe('GET /api/organizations/:id/campaigns', function () {
@@ -22,8 +53,7 @@ describe('Integration | Application | Route | campaign detail router', function 
       // given
       const method = 'GET';
       const url = '/api/organizations/1/campaigns';
-
-      const httpTestServer = new HttpTestServer();
+      httpTestServer = new HttpTestServer();
       await httpTestServer.register(moduleUnderTest);
 
       // when
@@ -32,16 +62,6 @@ describe('Integration | Application | Route | campaign detail router', function 
       // then
       expect(response.statusCode).to.equal(200);
       expect(campaignDetailController.findPaginatedFilteredCampaigns).to.have.been.calledOnce;
-    });
-  });
-
-  describe('GET /api/campaigns/{id}/csv-profiles-collection-results', function () {
-    it('should exist', async function () {
-      // when
-      const response = await httpTestServer.request('GET', '/api/campaigns/1/csv-profiles-collection-results');
-
-      // then
-      expect(response.statusCode).to.equal(200);
     });
   });
 });
