+++ api/tests/prescription/campaign/unit/domain/usecases/start-writing-campaign-assessment-results-to-stream_test.js
@@ -2,12 +2,12 @@ import stream from 'stream';
 
 const { PassThrough } = stream;
 
-import { expect, sinon, domainBuilder, streamToPromise, catchErr } from '../../../test-helper.js';
-import { startWritingCampaignAssessmentResultsToStream } from '../../../../lib/domain/usecases/start-writing-campaign-assessment-results-to-stream.js';
-import { UserNotAuthorizedToGetCampaignResultsError, CampaignTypeError } from '../../../../lib/domain/errors.js';
-import * as campaignCsvExportService from '../../../../lib/domain/services/campaign-csv-export-service.js';
-import { getI18n } from '../../../tooling/i18n/i18n.js';
-import { StageCollection } from '../../../../src/shared/domain/models/user-campaign-results/StageCollection.js';
+import { expect, sinon, domainBuilder, streamToPromise, catchErr } from '../../../../../test-helper.js';
+import { startWritingCampaignAssessmentResultsToStream } from '../../../../../../src/prescription/campaign/domain/usecases/start-writing-campaign-assessment-results-to-stream.js';
+import { UserNotAuthorizedToGetCampaignResultsError, CampaignTypeError } from '../../../../../../lib/domain/errors.js';
+import * as campaignCsvExportService from '../../../../../../src/prescription/campaign/domain/services/campaign-csv-export-service.js';
+import { getI18n } from '../../../../../tooling/i18n/i18n.js';
+import { StageCollection } from '../../../../../../src/shared/domain/models/user-campaign-results/StageCollection.js';
 
 describe('Unit | Domain | Use Cases | start-writing-campaign-assessment-results-to-stream', function () {
   const campaignRepository = { get: () => undefined };
